Extract FilterType interface
