+++ src/FilterType/DefaultFilterType.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Unlooped\GridBundle\FilterType;
+
+class DefaultFilterType extends AbstractFilterType
+{
+}
